✨(backend) add soft delete to documents and refactor db queryset
Now that we have introduced a document tree structure, it is not
possible to allow deleting documents anymore as it impacts the whole
subtree below the deleted document and the consequences are too big.

We introduce soft delete in order to give a second though to the
document's owner (who is the only one to be allowed to deleted a
document). After a document is soft deleted, the owner can still
see it in the trashbin (/api/v1.0/documents/?is_deleted=true).
After a grace period (30 days be default) the document disappears
from the trashbin and can't be restored anymore. Note that even
then it is still kept in database. Cleaning the database to erase
deleted document after the grace period can be done as a maintenance
script.